add unlock coin test
