+++ x/asset/handler_test.go
@@ -360,223 +360,3 @@ func TestIssueIsCanIssueTag(t *testing.T) {
 		simapp.CheckBalance(t, app, account4, amt)
 	})
 }
-
-func TestLockCoins(t *testing.T) {
-	app, _ := createAppForTest()
-
-	Convey("test lock core coins", t, func() {
-		ctx := app.NewTestContext()
-
-		lockedCoins := types.NewInt64Coins(constants.DefaultBondDenom, 1000000000)
-		lockedBlockNum := app.LastBlockHeight() + 1 + 5
-
-		msgLock := assetTypes.NewMsgLockCoin(addr4, account4, lockedCoins, lockedBlockNum)
-		tx := simapp.NewTxForTest(
-			account4,
-			[]sdk.Msg{
-				&msgLock,
-			}, wallet.PrivKey(addr4))
-
-		So(simapp.CheckTxs(t, app, ctx, tx), ShouldBeNil)
-
-		ctx = app.NewTestContext()
-		allLockedCoins, locks, err := app.AssetKeeper().GetLockCoins(ctx, account4)
-		So(err, ShouldBeNil)
-		So(allLockedCoins, simapp.ShouldEq, lockedCoins)
-		So(len(locks), ShouldEqual, 1)
-		So(locks[0].UnlockBlockHeight, ShouldEqual, lockedBlockNum)
-		So(locks[0].Coins, simapp.ShouldEq, lockedCoins)
-
-		currCoins := app.AssetKeeper().GetAllBalances(ctx, account4)
-		useableCoins := currCoins.Sub(lockedCoins).Sub(simapp.DefaultTestFee)
-
-		// when locked, coins can be transfer in useful coins
-		err = transfer(t, app, true, account4, account1, useableCoins, account4)
-		So(err, ShouldBeNil)
-
-		currCoinsBeforeFailedTransferCoins := app.AssetKeeper().GetAllBalances(app.NewTestContext(), account4)
-
-		// when locked, coins cannot be transfer
-		err = transfer(t, app, false, account4, account1, types.NewInt64Coins(constants.DefaultBondDenom, 100), account4)
-		So(err, simapp.ShouldErrIs, assetTypes.ErrAssetCoinsLocked)
-
-		currCoins = app.AssetKeeper().GetAllBalances(app.NewTestContext(), account4)
-		So(currCoinsBeforeFailedTransferCoins, simapp.ShouldEq, currCoins)
-	})
-
-	Convey("test no core coins lock", t, func() {
-		var (
-			symbol             = types.MustName("abc")
-			denom              = types.CoinDenom(name2, symbol) // create in last
-			maxSupplyAmt int64 = 10000000000000
-		)
-
-		So(createCoin(t, app, true, account2, symbol, maxSupplyAmt), ShouldBeNil)
-
-		issueAmt := types.NewInt64Coin(denom, 1000)
-		So(issueCoin(t, app, true, account2, symbol, issueAmt), ShouldBeNil)
-
-		var (
-			ctx            = app.NewTestContext()
-			lockedCoins    = types.NewInt64Coins(denom, 100)
-			lockedBlockNum = app.LastBlockHeight() + 1 + 5
-		)
-
-		msgLock := assetTypes.NewMsgLockCoin(addr2, account2, lockedCoins, lockedBlockNum)
-		tx := simapp.NewTxForTest(
-			account2,
-			[]sdk.Msg{
-				&msgLock,
-			}, wallet.PrivKey(addr2))
-
-		So(simapp.CheckTxs(t, app, ctx, tx), ShouldBeNil)
-
-		allLockedCoins, locks, err := app.AssetKeeper().GetLockCoins(app.NewTestContext(), account2)
-		So(err, ShouldBeNil)
-		So(allLockedCoins, simapp.ShouldEq, lockedCoins)
-		So(len(locks), ShouldEqual, 1)
-		So(locks[0].UnlockBlockHeight, ShouldEqual, lockedBlockNum)
-		So(locks[0].Coins, simapp.ShouldEq, lockedCoins)
-
-		// lock many coins
-		var (
-			lockedCoins2    = types.NewInt64Coins(constants.DefaultBondDenom, 100)
-			lockedBlockNum2 = app.LastBlockHeight() + 100
-		)
-
-		msgLock2 := assetTypes.NewMsgLockCoin(
-			addr2, account2, lockedCoins2, lockedBlockNum2)
-		tx2 := simapp.NewTxForTest(
-			account2,
-			[]sdk.Msg{
-				&msgLock2,
-			}, wallet.PrivKey(addr2))
-
-		So(simapp.CheckTxs(t, app, app.NewTestContext(), tx2), ShouldBeNil)
-
-		allLockedCoins, locks, err = app.AssetKeeper().GetLockCoins(app.NewTestContext(), account2)
-		So(err, ShouldBeNil)
-		So(allLockedCoins, simapp.ShouldEq, lockedCoins.Add(lockedCoins2...))
-		So(len(locks), ShouldEqual, 2)
-		So(locks[0].UnlockBlockHeight, ShouldEqual, lockedBlockNum)
-		So(locks[0].Coins, simapp.ShouldEq, lockedCoins)
-		So(locks[1].UnlockBlockHeight, ShouldEqual, lockedBlockNum2)
-		So(locks[1].Coins, simapp.ShouldEq, lockedCoins2)
-	})
-}
-
-func TestLockCoinsError(t *testing.T) {
-	app, _ := createAppForTest()
-
-	Convey("test lock err by denom coins no exist", t, func() {
-		var (
-			symbol             = types.MustName("abc")
-			denom              = types.CoinDenom(name2, symbol) // create in last
-			maxSupplyAmt int64 = 10000000000000
-		)
-
-		So(createCoin(t, app, true, account2, symbol, maxSupplyAmt), ShouldBeNil)
-
-		issueAmt := types.NewInt64Coin(denom, 1000)
-		So(issueCoin(t, app, true, account2, symbol, issueAmt), ShouldBeNil)
-
-		msgLock := assetTypes.NewMsgLockCoin(
-			addr2, account2, types.NewInt64Coins("aaa/coin", 100), 10)
-		tx := simapp.NewTxForTest(
-			account2,
-			[]sdk.Msg{
-				&msgLock,
-			}, wallet.PrivKey(addr2)).WithCannotPass()
-
-		So(simapp.CheckTxs(t, app, app.NewTestContext(), tx),
-			simapp.ShouldErrIs, assetTypes.ErrAssetCoinNoExit)
-
-	})
-
-	Convey("test lock err by no coins", t, func() {
-		var (
-			symbol             = types.MustName("aaa")
-			denom              = types.CoinDenom(name2, symbol) // create in last
-			maxSupplyAmt int64 = 10000000000000
-		)
-
-		So(createCoin(t, app, true, account2, symbol, maxSupplyAmt), ShouldBeNil)
-		// no issue, so account2 no have this coin
-
-		msgLock := assetTypes.NewMsgLockCoin(
-			addr2, account2, types.NewInt64Coins(denom, 100), 10)
-		tx := simapp.NewTxForTest(
-			account2,
-			[]sdk.Msg{
-				&msgLock,
-			}, wallet.PrivKey(addr2)).WithCannotPass()
-
-		So(simapp.CheckTxs(t, app, app.NewTestContext(), tx),
-			simapp.ShouldErrIs, assetTypes.ErrAssetLockCoinsNoEnough)
-	})
-
-	Convey("test lock err by block height err", t, func() {
-		var (
-			symbol             = types.MustName("aab")
-			denom              = types.CoinDenom(name2, symbol) // create in last
-			maxSupplyAmt int64 = 10000000000000
-		)
-
-		So(createCoin(t, app, true, account2, symbol, maxSupplyAmt), ShouldBeNil)
-		// no issue, so account2 no have this coin
-		issueAmt := types.NewInt64Coin(denom, 1000)
-		So(issueCoin(t, app, true, account2, symbol, issueAmt), ShouldBeNil)
-
-		simapp.AfterBlockCommitted(app, 2)
-
-		msgLock := assetTypes.NewMsgLockCoin(
-			addr2, account2, types.NewInt64Coins(denom, 100), 2) // 2 has already passed
-		tx := simapp.NewTxForTest(
-			account2,
-			[]sdk.Msg{
-				&msgLock,
-			}, wallet.PrivKey(addr2)).WithCannotPass()
-
-		So(simapp.CheckTxs(t, app, app.NewTestContext(), tx),
-			simapp.ShouldErrIs, assetTypes.ErrAssetLockUnlockBlockHeightErr)
-	})
-
-	Convey("test lock err by coins type cannot be locked", t, func() {
-		var (
-			symbol             = types.MustName("aac")
-			denom              = types.CoinDenom(name2, symbol) // create in last
-			maxSupplyAmt int64 = 10000000000000
-			maxSupply          = types.NewCoin(denom, types.NewInt(maxSupplyAmt))
-			initSupply         = types.NewCoin(denom, types.NewInt(0))
-			desc               = []byte(fmt.Sprintf("desc for %s", denom))
-		)
-
-		So(createCoinExt(t, app, true, account2, symbol, maxSupply, true, false, true, 0, initSupply, desc), ShouldBeNil)
-		// no issue, so account2 no have this coin
-		lockAmt := types.NewInt64Coin(denom, 1000)
-		So(issueCoin(t, app, true, account2, symbol, lockAmt), ShouldBeNil)
-
-		msgLock := assetTypes.NewMsgLockCoin(
-			addr2, account2, types.NewInt64Coins(denom, 100), 100) // 2 has already passed
-		tx := simapp.NewTxForTest(
-			account2,
-			[]sdk.Msg{
-				&msgLock,
-			}, wallet.PrivKey(addr2)).WithCannotPass()
-
-		So(simapp.CheckTxs(t, app, app.NewTestContext(), tx),
-			simapp.ShouldErrIs, assetTypes.ErrAssetCoinCannotBeLock)
-
-		// for multiple coins test
-		locks := types.NewInt64Coins(constants.DefaultBondDenom, 10).Add(lockAmt)
-		msgLock2 := assetTypes.NewMsgLockCoin(addr2, account2, locks, 100)
-		tx2 := simapp.NewTxForTest(
-			account2,
-			[]sdk.Msg{
-				&msgLock2,
-			}, wallet.PrivKey(addr2)).WithCannotPass()
-
-		So(simapp.CheckTxs(t, app, app.NewTestContext(), tx2),
-			simapp.ShouldErrIs, assetTypes.ErrAssetCoinCannotBeLock)
-	})
-}
